+++ tests/integration/test_exec.py
@@ -1,8 +1,7 @@
 __copyright__ = "Copyright (c) 2020-2021 Jina AI Limited. All rights reserved."
 __license__ = "Apache-2.0"
 
-from jina import Flow, Document, DocumentArray
-from jinahub.encoder.laser_encoder import LaserEncoder
+from jina import Flow, Document
 
 
 def data_generator(num_docs):
@@ -16,39 +15,6 @@ def test_use_in_flow():
     with Flow.load_config('flow.yml') as flow:
         data = flow.post(on='/encode', inputs=data_generator(5))
         docs = data[0].docs
+        assert len(docs) == 5
         for doc in docs:
             assert doc.embedding.shape == (1024,)
-
-
-def test_traversal_path():
-    text = 'blah'
-    docs = [Document(id='root1', text=text)]
-    docs[0].chunks = [Document(id='chunk11', text=text),
-                      Document(id='chunk12', text=text),
-                      Document(id='chunk13', text=text)
-                      ]
-    docs[0].chunks[0].chunks = [
-        Document(id='chunk111', text=text),
-        Document(id='chunk112', text=text),
-    ]
-
-    f = Flow().add(uses={
-        'jtype': LaserEncoder.__name__,
-        'with': {
-            'default_traversal_paths': ['c'],
-        }
-    })
-    with f:
-        result = f.post(on='/test', inputs=docs, return_results=True)
-        for path, count in [['r', 0], ['c', 3], ['cc', 0]]:
-            assert len(DocumentArray(result[0].data.docs).traverse_flat([path]).get_attributes('embedding')) == count
-
-        result = f.post(on='/test', inputs=docs, parameters={'traversal_paths': ['cc']}, return_results=True)
-        for path, count in [['r', 0], ['c', 0], ['cc', 2]]:
-            assert len(DocumentArray(result[0].data.docs).traverse_flat([path]).get_attributes('embedding')) == count
-
-
-def test_no_documents():
-    with Flow().add(uses=LaserEncoder) as f:
-        result = f.post(on='/test', inputs=[], return_results=True)
-        assert result[0].status.code == 0
