+++ tests/unit/test_exec.py
@@ -1,10 +1,9 @@
 __copyright__ = "Copyright (c) 2020-2021 Jina AI Limited. All rights reserved."
 __license__ = "Apache-2.0"
 
-from jina.executors import BaseExecutor
 import pytest
-from jinahub.encoder.laser_encoder import LaserEncoder
 from jina import Document, DocumentArray
+from jinahub.encoder.laser_encoder import LaserEncoder
 
 
 @pytest.fixture()
@@ -19,3 +18,33 @@ def test_flair_batch(docs_generator):
 
     assert len(docs.get_attributes('embedding')) == 30
     assert docs[0].embedding.shape == (1024,)
+
+
+def test_traversal_path():
+    text = 'blah'
+    docs = DocumentArray([Document(id='root1', text=text)])
+    docs[0].chunks = [Document(id='chunk11', text=text),
+                      Document(id='chunk12', text=text),
+                      Document(id='chunk13', text=text)
+                      ]
+    docs[0].chunks[0].chunks = [
+        Document(id='chunk111', text=text),
+        Document(id='chunk112', text=text),
+    ]
+
+    encoder = LaserEncoder()
+    encoder.encode(docs, parameters={'batch_size': 10, 'traversal_paths': ['c']})
+
+    for path, count in [['r', 0], ['c', 3], ['cc', 0]]:
+        assert len(docs.traverse_flat([path]).get_attributes('embedding')) == count
+
+    encoder.encode(docs, parameters={'batch_size': 10, 'traversal_paths': ['cc']})
+    for path, count in [['r', 0], ['c', 3], ['cc', 2]]:
+        assert len(docs.traverse_flat([path]).get_attributes('embedding')) == count
+
+
+def test_no_documents():
+    encoder = LaserEncoder()
+    docs = []
+    encoder.encode(docs, parameters={'batch_size': 10, 'traversal_paths': ['r']})
+    assert not docs
